Fix usage to health HIDL HAL.
- Before, if talking to the HIDL
HAL yields a transaction error (e.g. the HAL dies), because
of the lack of isOk(), perfetto crashes immediately. Now,
let GetBatteryCounter return false to indicate failure.

- Previously, the code also clears g_svc on Result::CALLBACK_DIED,
which never happens; this result only shows when dealing with
callbacks (when calling IHealth.update()).
I assume the original author wants to clear
the HAL pointer when the HAL dies. In this CL, I switch to a proper
DEAD_OBJECT check.

- Also fix a wrong type in a HIDL callback (int64_t->int32_t)

Test: builds
Bug: 208912315
Change-Id: I22de293d7465675928d31c879833b21f9cc41f8c