Avoid busy empty loop
In this test we have some attribute(constructor) functions. Each of them
checks for a specific environment variable to do something (if the
variable is not there, it returns immediately).

This is used in some testcases to launch a process that performs some
allocation in the following way:

* The test forks and exec itself with the extra env variable
* When the new process run, one attribute constructor function will
  detect the environment variable and perform some allocation.
* The new process shouldn't terminate: it should wait for the parent to
  kill it.
* An empty for loop achieves that.

It turns out that an empty busy loop is undefined behavior:

https://en.cppreference.com/w/cpp/language/memory_model#Progress_guarantee

and will trip clang 13+

The test code used to work on the android tree when built as part of the
`perfetto_integrationtests` binary because we passed the
`-fno-finite-loops` flag.

A future CL will need to enable the same code as part of some another
binary.

This commit adds an extra sleep(1) to workaround the problem and removes
`-fno-finite-loops`.

Bug: 195886381
Bug: 209967541
Bug: 206657946
Change-Id: I9a3e66b41001f323f682d3af1397bf00a9705f25