Do not use reachability to skip Cleaners for native size calculations
`reachable` in the `heap_graph_object` only considers strong references.
sun.misc.Cleaner is a phantom reference. We want to count a Cleaner if
it's still on the global `sun.misc.Cleaner.first` list: `reachable`
is not enough.

Now the code doesn't consider `reachable`, but looks at the `.next`
field: if it refers to itself, the Cleaner has been run (if null or if
it points to another object, the Cleaner has not been run).

Tested:
  `trace_processor_shell --run-metrics java_heap_stats` on the trace
  from b/209584534.

  Before aosp/1913867:
    heap_native_size: 761053420
    reachable_heap_native_size: 126174088
  Current master:
    heap_native_size: 114263016
    reachable_heap_native_size: 114263016
  After this change:
    heap_native_size: 287632108
    reachable_heap_native_size: 126174088

  Also, test/trace_processor/profiling/heap_graph_native_size.textproto
  covers this case already.

Bug: 213168086
Bug: 209584534
Change-Id: I67e044ec11b7710042e0ef3c6a08cfc7d2c3357c