Fix record_android_trace with config on android M and lower
adb shell on android M and lower does not support redirecting stdin from
host to device. stdin is used to send the config from the host to the
device in record_android_trace.

This commit works around the problem by copying the config on device
first and using `cat` to redirect it to the perfetto stdin.

Bug: 202539325
Change-Id: I9ab50d52632c02c49e1ebccf9950872c277ca6dd