Test: remove TreeHuggerOnly run e2e tests also on CI emulator
The rationale of TreehuggerOnly tests doesn't seem to hold
anymore. These tests seem to run fine on the Perfetto CI
emulator % some tweak to deal with a flakiness that shows
up only here.
Also cleanup a bit the usage of the macros, moving some of
the logic in the SetUp()


Bug: 73453011
Bug: 215221322
Change-Id: I68093151cd3d87a445f8816466e508ab3599464c