Add trusty_workqueues metric
Add trace-based metric that generates a simplified view of the Android
kworker events specific to the Trusty TEE. This generates a new track
per CPU that displays only the workqueue events related to Trusty. This
is a stopgap solution to aid in tracing issues around Trusty while the
Trusty team implements full tracing support.

Test: manual - trace processor and ui with custom trace
Bug: 211044277
Change-Id: Ic5ffd3b7b71ecf1babe648ae319a2aeff25acaee