[ui] Support nanosecond-scale zoom levels
Increases the maximum zoom and fixes rendering of nanosecond-level
events by disabling quantization for them and avoiding errors caused by
rounding.

Tested with a tool that's soon to be open source, which generates traces
using Intel Processor Trace of every function call/return in an execution.
Before this patch it isn't possible to zoom in all the way down to events
on the scale of single nanoseconds and with this patch it is.

Change-Id: I38eeeb9a63fd39d25c9fad05e1019a0a18ea8ad0