ui: Update ArgTree type check
When there is an argument with a 'length' field, the logic
incorrectly assumes it is an ArgTreeArray instead of an ArgTreeMap.

This leads to crash here:
http://shortn/_aUD7nsSK71

https://screenshot.googleplex.com/AY52WpejUPfwXdE

This solution changes the type check.

An alternative solution would have been to create `type` fields in
both `ArgTreeMap` and `ArgTreeArray` and then check on those.

Bug: 218298947
Change-Id: I0fe662811dd1539ed344c5bed0c0d642a42f2baf