GN: use arm64 on arm64 Macs
Rosetta-ness is inherited across processes. If we pull always
the x64 binary, then the output binaries on a M1 will be x64.
This is particularly bad because on x64 we assume SSE4.2 and
AVX2 but our binaries seem to crash with a SIGILL under Rosetta
when built with -mavx -mpopcnt -msse4.2.
Instead pull the arm64 GN binary, which should produce arm64
binaries.
Also:
- Fix the CPUID check, we should first check that XGETBV
  before invoking (we were doing that in the opposite order)
- Stop emitting "host_cpu=arm64" in build_all_configs. The
  default should JustWork(TM) now.

Bug: 218384527
Change-Id: I9365370d79bdf95dd7ff675172eb1194973c178b