Fix a bug related to kNoProgress status
GzipTraceParser, a client of GzipDecompressor was treating 'kNoProgress'
status as error and reporting "Failed to decompress". Which was not
really an error, instead it was just another form of 'kNeedsMoreInput'
(corresponding to Z_BUF_ERROR).

Luckily we never saw any practical impact of this bug because we always
fed mem-block of at least 1KB, so zlib always made some progress (i.e.
produced some output (mem-block of non-zero size)). This error can be
reproduced easily in a unit test which feeds mem-blocks of 3 bytes.

Also,

- Removed 'kNoProgress' code and merged it's usage with
  'kNeedsMoreInput', because there is no practical use of distinguishing
  between these two.
- Deleted the copy and move ops in GzipDecompressor, which was source
  of bug before this change. GzipDecompressor was default movable (
  because of unique_ptr), and the dtr of moved-from GzipDecompressor
  was accessing nullptr to call 'inflateEnd' method.
- Added in-comment comments (user-manual) in GzipDecompressor.

Change-Id: I0468dff47bd1831e13fb791ed92c2afa901749a9