Metrics: Introduce Jank Budget
In this CL, we introduce a new concept called JankBudget which is
defined as the minimum amount of time we need to reduce a frame
duration for it to be no longer be considered janky.

We also add total jank budget metric, which is calculated as the
the sum of jank budget of all janky frames to get an estimate of the
jank budget in the entire gesture (scroll, touch).

Why?
This will help us further improve our metric to pinpoint reasons and
cases with maximum impact in the reduction of jank and help
in prioritizing our efforts.

Doc explaining more about jank budget:
https://docs.google.com/document/d/1iYXaTQmOuJiWx_SXZlwaMRgTdq-0r7-6gcLxtqfe204/edit?usp=sharing

Bug: 209449166
Change-Id: Ic24c032365b4dc785c9bdf057c034649aff7c5ec