Make metric tag's text color unchanged when the track group is expanded
Previously the text color of the "metric" tag on the track group title
bar changes to white when the track group is expanded. This makes the
tag's text hard to read on its light blue background.

This CL resolves this issue by keeping the tag's text color unchanged
after track group expansion.

Bug: 220143471
Change-Id: I43effb822bdf109dc1cd309077926cd1e249e84f