Support InputMode::kDevNull in Subprocess
Subprocess always creates a pipe for stdin and passes it to the child
process.

This can sometimes be undesirable because (in a future commit) the child
process might not have SELinux permissions to read from a pipe.

This commit implements a new InputMode::kDevNull, which does not create
a pipe for stdin.

InputMode::kDevNull should always be used when args.input.empty(), but
existing callers might be subtly relying on the pipe, so this commit
doesn't change the default behavior.

Tested: unit tests.

**N.B.:** I haven't tested this on windows.

Bug: 219393750
Change-Id: Ifd900f8be595439d12d6632c26426291c9064001