heapprofd client api: cut unnecessary dependency on common/proc_utils
This collection of procfs-parsing utility functions is used by the
profiler binaries, not by the heap profiling client code
(heapprofd_client_api.so). Keeping the client api's dependencies strict
is particularly desireable because it's depended on by libart.so (for
managed heap profiling), which is now an apex module and therefore has
its own quirks of building. So the fewer files we have in the transitive
depedency tree, the better.

The non-android standalone client (see client_api_factory_standalone.cc)
uses a function from proc_utils, so that dependency is retained. This
isn't a big deal since the standalone client bundles the entirety of
heapprofd statically anyway.

Change-Id: I9f200b96aba1788b1f1875e48961962182257f72