Export perfetto trace proto java library
Android App Traceur needs to make use of several messages defined in the
perfetto trace proto. This is due to underlying changes in how Traceur
can get the list of available atrace categories. Traceur now queries
Perfetto to get the list of atrace categories in binary protobuf form
and parses that to list the atrace categories.

Bug: 219393750

Test: manual - imported the library into Traceur updated the code to
      query perfetto and parse the returned message. Verified the
      categroeis that were returned matched what was listed by running
      'atrace --list_categories'.
Change-Id: Iffbcad6f121088192efbc091b2cf96fa2990a6a4