Add a summary of java views to our breakdown metric.
Currently we tend to grab the first event which isn't super interesting.
So instead we manually curate a list of important view names to pick
above the first (and fall back on the first if we can't).

This was tested against 20 traces to see if it roughly categorized
things reasonably. We might need to break down this more, and we also
had plenty of chorographer frame handlers without any trace events or
only the power voter which implies we have some blind spots as well.

Bug: 209448519
Change-Id: Ic9a1340640efb15e3aa48df36b540d5d36d5a1df