deobfuscation: match dexfiles that have been extracted into memory
For managed app profiler samples, trace processor currently uses the
mapping name when figuring out which package's deobfuscation data to
use. This patch extends that logic to deal with paths to in-memory
dexfiles, as extracted by ART from the apk. Example:
  [anon:dalvik-classes.dex extracted in memory from
  /data/app/~~uUgHYtbjPNr2VFa3byIF4Q==/
  com.perfetto.example-aC94wTfXRC60l2HJU5YvjQ==/base.apk]

The name of the mapping is not an API, but should be stable in practice
(this is handled by art/libartbase/base/zip_archive.cc).

I admit that the whole deobfuscation matching approach is still hacky,
but this improves the logic until we properly rework it.

Bug: 226873402
Change-Id: I724444fadc830dcc4f3e1c94e146f06a043bf358