Generate separate alerts for simultaneous dropped frame events
Previously when multiple processes yield dropped frames with the same
timestamp, the dropped frame alert slice is only shown for one of the
processes, leading to underreporting of dropped frames from the other
processes. This CL now makes sure that alert slices are created for
all processes in the above situation.

This CL also renames the "process_id" in the DroppedFrame message to
the more conventional "pid".

Bug: 227660470
Change-Id: I131c8da95b36611cf29012290f7c56b1a2a68198