ftrace/cpu_reader: Avoid logging in fuzzer
Commit 2397d1a2d071("Dump malformed ftrace pages to log") added logs
when a ftrace page is malformed in CpuReader::ProcessPagesForDataSource.

Since CpuReader::ProcessPagesForDataSource is the entry point for the
fuzzer, these log messages also show up there, which is not a good
thing, because it slows down the fuzzer.

This commit moves the log messages to the caller of
CpuReader::ProcessPagesForDataSource and removes some redundant DCHECKs
(there already is a CHECK/FATAL in the caller).

Change-Id: If883f54bcd5571d09a8f2b8cb35e059d91706fda