Fix value_at_max_ts() for zero timestamps
value_at_max_ts ignored zero timestamps due to the 'maximum timestamp
seen' variable being initialised to zero. This caused an assertion in
the UI to fail in some situations. Fix this by initialising the variable
to min int instead and also making the comparison <= rather than = which
solves an analogous issue which could occur with min int.

Bug: 218895702
Change-Id: If54fe3b10786b6fb7152c7b854a0f45094c9a526