Filter debug annotations based on flag
Add support for a new flag in track_event configs in perfetto client
library to filter out debug annotations. This is only a subset of
fields we want to remove for privacy_filtering. The goal here to
prevent these two fields even while writing trace, to also reduce the
size of trace.

Bug: b/228292224
Change-Id: I9899ab66ab7f0931aba7f4ebf82eec5096b2b9e1