Change TraceWriter to support serialization from C ABI
This patch adds a way to send data to the various `TraceWriter`s without
using `protozero::Message`, by:

* Adding a MessageHandleBase::TakeStreamWriter() method, which allows
  direct access to the ScatteredStreamWriter, after calling
  TraceWriterBase::NewTracePacket().
* Adding a ScatteredStreamWriter::AnnotatePatch() method, which exposes
  protozero message length patching (and allows exporting that to the C
  ABI).
* Adding a TraceWriterBase::FinishTracePacket() method, which allows
  callers to commit a trace packet without relying on protozero::Message
  finalization.

With all these changes, it will be easy for a future commit, given the
pointer to a TraceWriter, to serialize TracePackets across the shared
library ABI.

`PatchList` used to assume a specific order of the offsets in the
patches, which was only feasible because TraceWriterImpl created patches
from the outermost message to the most nested. This can no longer be
assumed now, because callers can call AnnotatePatch() in the order they
prefer.

Bug: 237053538
Change-Id: I652777daf599875bd3531c49e4899dfdf1282609