tools: Move --serial to base argument parser
Currently --serial is listed under the arguments group titled
"Full trace config (only when not using short options)". This makes
it sound like like --serial is unable to be used with the "light"
configuration options, which is not the case.

--serial simply sets the ANDROID_SERIAL env variable, which is
applicable for both light and full options, so move this out of
the "Full trace config" options group to make it clear to the
user that they can also use it with the "light" config when
they run "perfetto --help"

testing:

run "./tools/record_android_trace --help" and verify the --serial
option is printed with the base options.

Change-Id: I9c5b0e6bc830b33a5d92e5d4e3d59560bf692d77