Show absolute time in UI, adds ABS_TIME_STR to SQL
Shows the absolute time in the details panel via an ABS_TIME_STR
extension used to get a formatted time with nanosecond precision.
It returns a formatted time instead of unix epoch nanoseconds
to avoid the JS UI rounding large numbers by converting to doubles.

Based on aosp/1965224 which can't be updated as the user account
was deleted.

Change-Id: Ie843bc92a0f79cad90a7d9c9867897a0c4a78f52