perf: support for kthreads and kernel-only callstacks
Before this patch, traced_perf always dropped samples corresponding to
kernel threads, as a leftover of the times when we only supported
userspace callstacks. After this patch, we keep the kthread samples if
kernel callstacks are enabled. Additionally, we now support kernel-only
callstack sampling. This will keep any kthread samples, as well as any
kernel frames if a sampled userspace process happens to be in kernel
context.

Kthreads don't have a /proc/pid/cmdline, but we still apply any
|pid|-level deny/allowlists from the config.

The proto options for userspace callstacks is an enum because I can
foresee a future where we enable alternative unwinding strategies (such
as asking the kernel to unwind based on framepointers) on appropriate
platforms. The "kernel_frames" option is still a bool since there are
really only two possibilities there.

Bug: 227649697
Change-Id: I026ee01659a9c1868bd794c0d6683e9d50ac3ad8