tp: associate reserved utid0 to the reserved upid0
This affects Linux traces, which use these identifiers for the idle
process. The net effect is that any thread and process scoped tracks
for swapper will stay grouped in the UI. Concrete example: kernel
systrace events from interrupt context. B/E events that are
thread-scoped, and C events that are process-scoped as of aosp/2076218.

Change-Id: I35f72a810eb849113806a6f5adf3bd5c90b494f5