(v2) tp: associate reserved utid0 to the reserved upid0
This affects Linux traces, which use these identifiers for the idle
process. The net effect is that any thread and process scoped tracks
for swapper will stay grouped in the UI. Concrete example: kernel
systrace events from interrupt context. B/E events that are
thread-scoped, and C events that are process-scoped as of aosp/2076218.

Now with a "utid != 0" case for track_decider so that the sched data
attached to utid 0 doesn't push that track to the top of the UI (based
on runtime).

Change-Id: I890e2d36eee41697a2cc4cdd9ed869e78512f413