ui: Add eslint
tslint is deprecated (more over it's started printing worrying errors
when running). The suggested upgrade path is eslint. This installs
eslint and adds the Google style guide rules. Currently
  git cl format --js
and
  eslint
disagree on the indentation so that will have to be resolved before we
switch over to eslint entirely.

I ran:
ui/eslint --fix src/frontend/aggregation_panel.ts
so the diff in src/frontend/aggregation_panel.ts is an an example of
some of the issues eslint found in the code base.

Change-Id: I3584f2701b89172d0b13500edfa124b1324c0381