Update googletest version for standalone build
The new version of googletest brings some benefits: for example it
detects when a TEST_P doesn't have a corresponding
INSTANTIATE_TEST_SUITE_P, which is usually an error. Otherwise, the
problem will only be detected later in the chromium/android builds.

The new version is the latest version of the github sources that Android
has merged. It is older than the one chromium is currently using.

Change-Id: I23f8ce9e7e4177026c500977db7b881ed5fbc4b4