[ui] Support showing results of the multiple queries
Support showing results of multiple queries at the same time to support
more complex investigations and metric development.

Now if Ctrl/Cmd+Enter is pressed in the omnibox-query mode (as opposed
to Enter), the results of the query will be displayed in a separate
panel.

Also the results of a query from a standalone page will be also
displayed in the timeline mode.

To distinguish between them, the tab names are renamed from 'Query
Result' to 'Omnibox Query', 'Standalone Query' and 'Pinned Query'
respectively.

R=ddrone@google.com
BUG=b/233383594

Change-Id: I01ddd76a052e5f29e2234bfc077d29d0358d2786