ui: Add custom signing with RSA key
Create a (private, public) key pair used for authenticating with the
device.

Add the jsbn-rsa package for creating an RSA key.

We will sign with a standard RSA private key but we will transform
the token we receive from adb prior.
For the public key, we will customise it to work with adb.

Change-Id: Ia5d8e28bfc9a044bb45480d1e86ec0b58d3101af