tp: improve translation capabilities for slice args
This CL makes it possible to translate slice args *after* the slice
is seen instead of just before like we supported for now.

Inspiration for this CL was taken from aosp/2102099 but with a
simpler structure (introducing less new code and integrating better
with existing TP concepts).

See aosp/2102099 for context on why this CL is needed.

Bug: 220179588
Change-Id: I6177364b534750debf0316ac8a07236642f1dd49