[fuchsia] Reland: Add alternate socket acceptance path.
IPC interface is slightly altered in a way to allow embedders to
create Hosts which are unbacked by listen sockets.
Connected sockets are then supplied to these non-listening Hosts
by the embedder by invoking a new method,
Host::AdoptConnectedSocket_Fuchsia().

This is an enabling set of changes for enabling Perfetto to be
used as a system tracing service on Fuchsia.

* Tests updated to make use of AdoptConnectedSocket-based path
  on Fuchsia.
* Some testing IPC defaults revisited for Fuchsia.
* Misc. cleanup: class member naming (missing underscores),
     typos in comments, etc.

(Reland CL: fixes some CL edits that were not exercised under
Fuchsia prior to patch upload due to human error. Also omits
SOCK_CLOEXEC protocol, which isn't an allowed flag for AF_UNIX
socketpairs on Fuchsia since it does not employ the UNIX exec()
model which FD_CLOEXEC affects.)

Test: perfetto_unittests pass on Linux, Fuchsia
Bug: chromium:1328930
Change-Id: I700b76c226379371799b13f5ee63d1c78c67b2f2