[fuchsia] Add alternate socket acceptance path.
IPC interface is slightly altered in a way to allow embedders to
create Hosts which are unbacked by listen sockets.
Connected sockets are then supplied to these non-listening Hosts
by the embedder by invoking a new method,
Host::AdoptConnectedSocket_Fuchsia().

This is an enabling set of changes for enabling Perfetto to be
used as a system tracing service on Fuchsia.

* Tests updated to make use of AdoptConnectedSocket-based path
  on Fuchsia.
* Some testing IPC defaults revisited for Fuchsia.
* Misc. cleanup: class member naming (missing underscores),
     typos in comments, etc.

Test: perfetto_unittests pass on Linux, Fuchsia
Bug: chromium:1328930

Change-Id: I0426f488ec15b630c827154c967c68e4b7638b11