Remove path from process name in Long Latency alert
When a process name containing pathname to a file is displayed as the
title of a track group, Perfetto removes the path from the process name.
However, the Long Latency alert still shows the process name with the
original path, causing inconsistency.

This CL removes the path part of the process name shown in the Long
Latency alert.

Bug: 235140211
Change-Id: I681d0410a9f1e22a02b4b76c5c721b3180382a9e