Add ability to aggregate by MIN and MAX in Pivot Table
This CL adds a popup menu on aggregation columns, allowing the user to
change the aggregation function applied to the column.

Screenshot: https://screenshot.googleplex.com/3UFdnQbZZC6G2gi

Known issue: changing the aggregation function puts the column as the
last one, this is going to be addressed separately as a part of CL
allowing to change the order of columns via drag'n'drop.

Bug: 232769552
Change-Id: I9ba479caa98c61dd1d085acb1b5b4339c387c7d2