Fix concurrency of GetDataSourceLocked() and  GetInterceptorLocked()
In GetDataSourceLocked, internal_state->data_source is a unique pointer,
which is supposed to be protected by internal_state->lock. Due to a
limitation of the LockedHandle constructor, the
internal_state->data_source unique pointer was accessed before acquiring
the lock.

The same used to happen in GetInterceptorLocked.

This commit fixes the problem by acquiring the lock earlier.

The problem was found by reading the code.

Change-Id: I21682015bf21fbcad39f8cc425804550d2a57b7a