[ui] Fix process sorting by perf samples
"Select distinct" does not have the desired behavior when there is
more than one process present with perf samples. The existing query
will elevate on of those processes to the top of the process list,
but leave the other processes in place.

This fixes the process sorting when multiple processes have perf
samples.

Testing: ./tools/cpu_profile -n <app1>,<app2>

verify when opened in the UI that BOTH app1 and app2 appear at the
top of the trace.

Change-Id: I03f70c850f304196b42ce4fe2e245d4b9d40cad1