gn: Minimize rebuilding when commit hash changes
Today, src/base:base depends on a generated header that embeds the
perfetto version (therefore the repository commit hash).

Since almost everything depends on src/base:base, when the commit hash
changes, everything is rebuilt.

This is kind of wasteful when changing commits often (e.g., when doing
bisects).

This commit fixes the problem by putting version header into its own gn
target, which is going to be depended on by only a few other targets.

This avoids, for example, rebuilding the protoc_plugin, which triggers
the recompilation on most of the perfetto code.

Change-Id: I68140570e1abda489d9f30582eed5f8ddb7b3ead