[ui] optimize sched subquery in process sorting
Apply similar optimizations for querying CPU duration as was
implemented for querying per-process sclice count.

This optimization relies on the the fact that in SQLite joins perform
much better when they are joining less data. Here in order reduce
the data we need to join, we first aggregate sched duration by utid
prior to joining.

This optimiztion causes the subquery to perform 2x as fast on the
example_android_trace_15s sample trace provided by the UI and the
full query ~50% faster.

Change-Id: Ib06c959ac72b0dc5cb554c7247705ef4b0fa280c