ui: Change interface to support datasources for targets
Define a DataSource to have a name and an `unknown` which will
contain the information. This field will be opaque to the recording
code, but the caller can cast it using the name.

Remove DynamicTargetInfo because now there are multiple pieces
of information that we can get dynamically from the device and
the nesting under a single field is too restrictive.

Bug: 209947455
Change-Id: I6e0bb7ea15d2661616ea70a671f191913e842c96