ui: optimize perf sample sorting sub-query
Similar to how the subqueries for slice count and sched were optimized
we also similarly optimize the sub-query for sorting by perf sample
count.

This is done by aggregating sample counts per-utid before the join
operation happens. This change also removes the join with the process
table, which isn't needed as the thread table already has the
correctly mapped upid

Change-Id: I1fc6b2e161b5cf0f5135dd0b9cdf2d4054d74a12