Start phasing out checkbox UI for Pivot Table configuration
Current UI for Pivot Table was intended to be temporary one. This CL
starts the process of phasing it out. Specifically, the pivot table
will start from table view in default configuration; and instead of
"Edit" button taking a full strip of vertical space it's moved to a
popup menu in the top-right corner.

Screenshot: https://screenshot.googleplex.com/BbnNJocqmEZjviT

Change-Id: I3e2f4e39e6115b415d5b7274d1833bdce11045c9