traced_probes: Add statsd atoms datasource
Currently this only works when side loaded as it execs 'cmd stats'
which we can do as shell but not with in tree builds due to selinux. There are too many selinux denials to reasonably prepose so in a
follow up I'll add support for talking binder directly to statsd.

Bug: 236596172
Change-Id: I0df694291f3fa7bc584d6623982cfba378feda3f