Don't emit consecutive empty trace packets
Empty packets are sometimes emitted by Chrome to ensure that the service
can safely read the last event from the trace buffer. But it doesn't
make sense to emit multiple empty packets in a row, and it may create
unnecessary overhead. So in this CL we skip all empty packets after the
first one.

This will ensure that we don't re-introduce regressions like this one:
crbug.com/1250114 when we switch Chrome to Perfetto client library.

Bug: b/201800654
Change-Id: Iccd5640babb945ddf56dc64c3ba937acd593040e