Show symbolization error for /data/local/tmp binaries
Default SELinux rules in android prevent profilers (heapprofd and
traced_perf) from symbolizing binaries located in /data/local/tmp.

If we get a call stack (for a heap profile or a cpu profile) from a
binary in /data/local/tmp with an empty build id (a sign that
symbolization failed), display an error message.

Bug: 238457976
Change-Id: I1122e4a548c4d5d209d25b6bd6e6d638bf36f230