Pivot table: get rid of edit view
The only remaining functionality of edit mode that was not available
in the dropdown configuration was a toggle that switches between
querying the whole timeline or just selected area. This CL removes the
edit mode and replaces popup menu item opening edit mode to the item
that toggles the data being queried.

Change-Id: I8d72787ed41083a70a88d5fc1ea4ce56c969466c