[ui] Add --cross-origin-isolation flag to ui/run-dev-server
Add a flag to the run-dev-server to start serving cross-origin isolation headers:
- Cross-Origin-Opener-Policy: same-origin
- Cross-Origin-Embedder-Policy: require-corp

These headers enable cross-origin isolation for the page (i.e.
window.crossOriginIsolated will be true) and as a consequence enable
high-precision timers for the web page.

The downside is that COOP: same-origin breaks openers and prevents
trace injection via postMessage, so for now this flag is test-only.

R=primiano@google.com

Change-Id: If037b79bb8df00d87b30c9188efdaf5bfe65acbf