Tests: deflake protozero_benchmark.cc
A code path was trying to fool the compiler and prove that
the result of some computation is used.
Unfortunately was doing so with a PERFETTO_CHECK.
Turns out in some cases we can end up hitting the == 42
branch because we indirectly hash the `state`.
Just replace that with a dummy log statement, which has
still side effects but doesn't break everything.

Flake: https://ci.perfetto.dev/#!/logs/20220726191943--cls-2166363-1--linux-gcc7-x86_64-release
Change-Id: Idaf579c1b0fb1690a119ab1fe132c23a843d02e4