Set correct track_uuid for legacy events
Before this change, we were not setting correct track_uuid for the
legacy events of TrackEvent::TYPE_UNSPECIFIED type.

For such events we need to set `track_event.track_uuid` to zero to
dissociate it with the default track.

Note: This CL may not be the complete fix for setting track_uuid
correctly in all permutations. For example, when the legacy event type
is not `TYPE_UNSPECIFIED`, there are certain cases where explicit
track_uuid is required but may not be correctly set as of now. This CL
doesn't address those cases. It only ensures to set track_uuid to zero
in case of `TYPE_UNSPECIFIED` events.

Change-Id: I7c5e0431dd034e4ff439b270bc7864e7a75052c7