Do not show Slices aggregation panel with Pivot Table
Pivot Table is supposed to solve the same problem slice aggregation
currently solves, in a more generic way. Therefore, when the Pivot Table
flag is turned on, there is no reason to show Slices panel as well.

This CL does exactly that, omitting Slices aggregation controller when
Pivot Table flag is turned on.

Bug: b/242535955
Change-Id: Ib6d2fa0b3fa583b5d08052969020b16f907c8c62